+++ examples/c/vector_add.py
@@ -3,26 +3,26 @@
 
 import numpy
 from kernel_tuner import tune_kernel
+from collections import OrderedDict
 
 kernel_string = """ 
 #include <omp.h>
-#include "timer.h"
 
 typedef float vfloat __attribute__ ((vector_size (vecsize*4)));
 
-float vector_add(vfloat *c, vfloat *a, vfloat *b, int n) {
-    unsigned long long start = get_clock();
-    int chunk = n/vecsize/nthreads;
+extern "C" float vector_add(vfloat *c, vfloat *a, vfloat *b, int n) {
+    double start = omp_get_wtime();
+    int chunk = n/(vecsize*nthreads);
 
     #pragma omp parallel num_threads(nthreads)
     {
         int offset = omp_get_thread_num()*chunk;
-        for (int i = offset; i<offset+chunk; i++) {
+        for (int i = offset; i<offset+chunk && i<n; i++) {
             c[i] = a[i] + b[i];
         }
     }
 
-    return (get_clock()-start) / get_frequency() / 1000000.0;
+    return (float)((omp_get_wtime() - start)*1e3);
 }
 """
 
@@ -35,10 +35,11 @@ n = numpy.int32(size)
 
 args = [c, a, b, n]
 
-tune_params = dict()
-tune_params["vecsize"] = [2**i for i in range(8)]
-tune_params["nthreads"] = [1, 2, 3, 4, 6, 8, 12, 16, 24, 32]
+tune_params = OrderedDict()
+tune_params["nthreads"] = [1, 2, 3, 4, 8, 12, 16, 24, 32]
+tune_params["vecsize"] = [1, 2, 4, 8, 16]
 
 answer = [a+b, None, None]
 
-tune_kernel("vector_add", kernel_string, size, args, tune_params, answer=answer)
+tune_kernel("vector_add", kernel_string, size, args, tune_params,
+    answer=answer, compiler_options=['-O3'])